+++ .gitignore
@@ -7,15 +7,15 @@ build/
 /captures
 .externalNativeBuild
 .cxx
-iosApp/Podfile.lock
-iosApp/Pods/*
-iosApp/iosApp.xcworkspace/*
-iosApp/iosApp.xcodeproj/*
-!iosApp/iosApp.xcodeproj/project.pbxproj
+sample/iosApp/Podfile.lock
+sample/iosApp/Pods/*
+sample/iosApp/iosApp.xcworkspace/*
+sample/iosApp/iosApp.xcodeproj/*
+!sample/iosApp/iosApp.xcodeproj/project.pbxproj
 webview/webview.podspec
 /convention-plugins/
 keySec.gpg
-desktopApp/jcef-bundle
-desktopApp/cache
-desktopApp/DawnCache
-desktopApp/GPUCache
+sample/desktopApp/jcef-bundle
+sample/desktopApp/cache
+sample/desktopApp/DawnCache
+sample/desktopApp/GPUCache
