+++ .run/desktopApp.run.xml
@@ -2,7 +2,7 @@
   <configuration default="false" name="desktopApp" type="GradleRunConfiguration" factoryName="Gradle">
     <ExternalSystemSettings>
       <option name="executionName" />
-      <option name="externalProjectPath" value="$PROJECT_DIR$/desktopApp" />
+      <option name="externalProjectPath" value="$PROJECT_DIR$/sample/desktopApp" />
       <option name="externalSystemIdString" value="GRADLE" />
       <option name="scriptParameters" value="" />
       <option name="taskDescriptions">
