--- desktopApp/src/jvmMain/kotlin/main.kt
@@ -1,61 +1,0 @@
-import androidx.compose.material.Text
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.window.Window
-import androidx.compose.ui.window.application
-import com.multiplatform.webview.MainWebView
-import com.multiplatform.webview.web.Cef
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import java.io.File
-import kotlin.math.max
-
-fun main() = application {
-    Window(onCloseRequest = ::exitApplication) {
-        var restartRequired by remember { mutableStateOf(false) }
-        var downloading by remember { mutableStateOf(0F) }
-        var initialized by remember { mutableStateOf(false) }
-
-        LaunchedEffect(Unit) {
-            withContext(Dispatchers.IO) {
-                Cef.init(builder = {
-                    installDir = File("jcef-bundle")
-                    settings {
-                        cachePath = File("cache").absolutePath
-                    }
-                }, initProgress = {
-                    onDownloading {
-                        downloading = max(it, 0F)
-                    }
-                    onInitialized {
-                        initialized = true
-                    }
-                }, onError = {
-                    it.printStackTrace()
-                }, onRestartRequired = {
-                    restartRequired = true
-                })
-            }
-        }
-
-        if (restartRequired) {
-            Text(text = "Restart required.")
-        } else {
-            if (initialized) {
-                MainWebView()
-            } else {
-                Text(text = "Downloading $downloading%")
-            }
-        }
-
-        DisposableEffect(Unit) {
-            onDispose {
-                Cef.dispose()
-            }
-        }
-    }
-}
