+++ sample/androidApp/src/androidMain/AndroidManifest.xml
