+++ sample/androidApp/src/androidMain/kotlin/com/multiplatform/webview/MainActivity.kt
@@ -1,0 +1,16 @@
+package com.multiplatform.webview
+
+import com.kevinnzou.sample.MainWebView
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.appcompat.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        setContent {
+            MainWebView()
+        }
+    }
+}
