+++ sample/androidApp/src/androidMain/res/drawable/ic_launcher_background.xml
