+++ sample/androidApp/src/androidMain/res/mipmap-hdpi/ic_launcher_round.png
