+++ sample/androidApp/src/androidMain/res/mipmap-mdpi/ic_launcher.png
