+++ sample/androidApp/src/androidMain/res/mipmap-xhdpi/ic_launcher_round.png
