+++ sample/androidApp/src/androidMain/res/mipmap-xxhdpi/ic_launcher_round.png
