+++ sample/androidApp/src/androidMain/res/values/strings.xml
