+++ sample/desktopApp/build.gradle.kts
@@ -1,0 +1,41 @@
+import org.jetbrains.compose.desktop.application.dsl.TargetFormat
+
+plugins {
+    kotlin("multiplatform")
+    id("org.jetbrains.compose")
+}
+
+kotlin {
+    jvm()
+    sourceSets {
+        val jvmMain by getting  {
+            dependencies {
+                implementation(compose.desktop.currentOs)
+                implementation(project(":sample:shared"))
+            }
+        }
+    }
+}
+
+compose.desktop {
+    application {
+        mainClass = "MainKt"
+
+        nativeDistributions {
+            targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
+            packageName = "KotlinMultiplatformComposeDesktopApplication"
+            packageVersion = "1.0.0"
+            includeAllModules = true
+        }
+    }
+}
+
+afterEvaluate {
+    tasks.withType<JavaExec> {
+        if (System.getProperty("os.name").contains("Mac")) {
+            jvmArgs("--add-opens", "java.desktop/sun.awt=ALL-UNNAMED")
+            jvmArgs("--add-opens", "java.desktop/sun.lwawt=ALL-UNNAMED")
+            jvmArgs("--add-opens", "java.desktop/sun.lwawt.macosx=ALL-UNNAMED")
+        }
+    }
+}
