+++ sample/iosApp/Configuration/Config.xcconfig
