+++ sample/iosApp/iosApp/Assets.xcassets/AccentColor.colorset/Contents.json
