+++ sample/iosApp/iosApp/Assets.xcassets/AppIcon.appiconset/Contents.json
