+++ sample/iosApp/iosApp/Assets.xcassets/AppIcon.appiconset/app-icon-1024.png
