+++ sample/iosApp/iosApp/Assets.xcassets/Contents.json
