+++ sample/iosApp/iosApp/Info.plist
