+++ sample/iosApp/iosApp/Preview Content/Preview Assets.xcassets/Contents.json
