+++ sample/shared/src/androidMain/kotlin/com/kevinnzou/sample/main.android.kt
@@ -1,0 +1,8 @@
+package com.kevinnzou.sample
+
+import androidx.compose.runtime.Composable
+
+actual fun getPlatformName(): String = "Android"
+
+@Composable
+fun MainWebView() = WebViewApp()
