+++ sample/shared/src/commonMain/kotlin/com/kevinnzou/sample/WebViewApp.kt
@@ -1,0 +1,46 @@
+package com.kevinnzou.sample
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.multiplatform.webview.web.WebView
+import com.multiplatform.webview.web.rememberWebViewState
+
+@Composable
+internal fun WebViewApp() {
+//    WebViewSample()
+    BasicWebViewSample()
+//    BasicWebViewWithHTMLSample()
+}
+
+@Composable
+internal fun WebViewSample() {
+    MaterialTheme {
+        val webViewState =
+            rememberWebViewState("https://github.com/KevinnZou/compose-webview-multiplatform")
+        webViewState.webSettings.apply {
+            isJavaScriptEnabled = true
+            customUserAgentString = "Mozilla/5.0 (Macintosh; Intel Mac OS X 11_1) AppleWebKit/625.20 (KHTML, like Gecko) Version/14.3.43 Safari/625.20"
+            androidWebSettings.apply {
+                isAlgorithmicDarkeningAllowed = true
+                safeBrowsingEnabled = true
+            }
+        }
+        Column(Modifier.fillMaxSize()) {
+            val text = webViewState.let {
+                "${it.pageTitle ?: ""} ${it.loadingState} ${it.lastLoadedUrl ?: ""}"
+            }
+            Text(text)
+            WebView(
+                state = webViewState,
+                modifier = Modifier.fillMaxSize()
+            )
+        }
+
+    }
+}
+
+expect fun getPlatformName(): String
