+++ sample/shared/src/iosMain/kotlin/com/kevinnzou/sample/main.ios.kt
@@ -1,0 +1,7 @@
+package com.kevinnzou.sample
+
+import androidx.compose.ui.window.ComposeUIViewController
+
+actual fun getPlatformName(): String = "iOS"
+
+fun MainViewController() = ComposeUIViewController { WebViewApp() }
