+++ webview/build.gradle.kts
@@ -43,8 +43,6 @@ kotlin {
         val androidMain by getting {
             dependencies {
                 api("androidx.activity:activity-compose:1.7.2")
-                api("androidx.appcompat:appcompat:1.6.1")
-                api("androidx.core:core-ktx:1.12.0")
                 api("androidx.webkit:webkit:1.8.0")
                 implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.3")
             }
