--- webview/src/androidMain/kotlin/com/multiplatform/webview/main.android.kt
@@ -1,8 +1,0 @@
-package com.multiplatform.webview
-
-import androidx.compose.runtime.Composable
-
-actual fun getPlatformName(): String = "Android"
-
-@Composable
-fun MainWebView() = WebViewApp()
