--- webview/src/desktopMain/kotlin/com/multiplatform/webview/main.desktop.kt
@@ -1,15 +1,0 @@
-package com.multiplatform.webview
-
-import androidx.compose.desktop.ui.tooling.preview.Preview
-import androidx.compose.runtime.Composable
-
-actual fun getPlatformName(): String = "Desktop"
-
-@Composable
-fun MainWebView() = WebViewApp()
-
-@Preview
-@Composable
-fun AppPreview() {
-    WebViewApp()
-}
