--- webview/src/iosMain/kotlin/com/multiplatform/webview/main.ios.kt
@@ -1,7 +1,0 @@
-package com.multiplatform.webview
-
-import androidx.compose.ui.window.ComposeUIViewController
-
-actual fun getPlatformName(): String = "iOS"
-
-fun MainViewController() = ComposeUIViewController { WebViewApp() }
